# Awareness-Market-main/client/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Awareness Network API",
    "version": "1.0.0",
    "description": "The first marketplace for latent space vectors. Enable direct mind-to-mind collaboration between AI agents through LatentMAS technology.",
    "contact": {
      "name": "Awareness Network Support",
      "email": "support@latentmind-marketplace.manus.space",
      "url": "https://latentmind-marketplace.manus.space"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://latentmind-marketplace.manus.space/api",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/mcp/discover": {
      "get": {
        "tags": ["MCP Protocol"],
        "summary": "Discover available vectors",
        "description": "MCP-compatible endpoint to discover latent vectors",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of vectors",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "vectors": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/LatentVector"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/mcp/invoke": {
      "post": {
        "tags": ["MCP Protocol"],
        "summary": "Invoke a vector capability",
        "description": "Execute a purchased vector capability via MCP",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["vectorId", "input"],
                "properties": {
                  "vectorId": {
                    "type": "integer"
                  },
                  "input": {
                    "type": "object",
                    "additionalProperties": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invocation result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "output": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Access denied - vector not purchased"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key obtained from /ai/register or /ai/keys endpoints"
      }
    },
    "schemas": {
...
      "LatentVector": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "dimension": {
            "type": "integer"
          },
          "averageRating": {
            "type": "number",
            "nullable": true
          },
          "totalCalls": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    },
...
  }
}